+++ ArsonistMod/Modules/Config.cs
@@ -43,6 +43,8 @@ namespace ArsonistMod.Modules
         public static ConfigEntry<bool> enableNonAggressiveHeatHaze;
 
         public static ConfigEntry<float> arsonistVoicelineVolume;
+        public static ConfigEntry<float> arsonistVoicelineVolumeArsonist;
+        public static ConfigEntry<float> arsonistVoicelineVolumeFirebug;
 
         /*
          
@@ -245,9 +247,23 @@ namespace ArsonistMod.Modules
 
             arsonistVoicelineVolume = ArsonistPlugin.instance.Config.Bind<float>
             (
+                new ConfigDefinition("03 - Voice", "Master Voice Volume"),
+                100f,
+                new ConfigDescription("Determines the volume for All voice lines, both Arsonist and Firebug.")
+            );
+
+            arsonistVoicelineVolumeArsonist = ArsonistPlugin.instance.Config.Bind<float>
+            (
                 new ConfigDefinition("03 - Voice", "Arsonist Voice Volume"),
                 100f,
-                new ConfigDescription("Determines the volume for Arsonist voice lines.")
+                new ConfigDescription("Determines the volume for Arsonist's voice lines. (Does not affect Firebug, affected by Master Voice Volume)")
+            );
+
+            arsonistVoicelineVolumeFirebug = ArsonistPlugin.instance.Config.Bind<float>
+            (
+                new ConfigDefinition("03 - Voice", "Firebug Voice Volume"),
+                100f,
+                new ConfigDescription("Determines the volume for Firebug's voice lines. (Does not affect Arsonist, affected by Master Voice Volume)")
             );
 
             baseGaugeLowerBoundRecharge = ArsonistPlugin.instance.Config.Bind<float>
@@ -386,6 +402,28 @@ namespace ArsonistMod.Modules
                     }
                     )
                 );
+            ModSettingsManager.AddOption(
+                new StepSliderOption(
+                    arsonistVoicelineVolumeArsonist,
+                    new StepSliderConfig
+                    {
+                        min = 0f,
+                        max = 100f,
+                        increment = 0.5f,
+                    }
+                    )
+                );
+            ModSettingsManager.AddOption(
+                new StepSliderOption(
+                    arsonistVoicelineVolumeFirebug,
+                    new StepSliderConfig
+                    {
+                        min = 0f,
+                        max = 100f,
+                        increment = 0.5f,
+                    }
+                    )
+                );
             ModSettingsManager.AddOption(new CheckBoxOption(shouldEnableDieKey));
 
             ModSettingsManager.AddOption(new CheckBoxOption(overheatTextShouldVibrate));
@@ -467,6 +505,24 @@ namespace ArsonistMod.Modules
         public static void OnChangeHooks() 
         {
             arsonistVoicelineVolume.SettingChanged += ArsonistVoicelineVolume_Changed;
+            arsonistVoicelineVolumeArsonist.SettingChanged += ArsonistVoicelineVolumeArsonist_SettingChanged; ;
+            arsonistVoicelineVolumeFirebug.SettingChanged += ArsonistVoicelineVolumeFirebug_SettingChanged; ;
+        }
+
+        private static void ArsonistVoicelineVolumeFirebug_SettingChanged(object sender, EventArgs e)
+        {
+            if (AkSoundEngine.IsInitialized())
+            {
+                AkSoundEngine.SetRTPCValue("Volume_ArsonistVoice_Firebug", arsonistVoicelineVolumeFirebug.Value);
+            }
+        }
+
+        private static void ArsonistVoicelineVolumeArsonist_SettingChanged(object sender, EventArgs e)
+        {
+            if (AkSoundEngine.IsInitialized())
+            {
+                AkSoundEngine.SetRTPCValue("Volume_ArsonistVoice_Arsonist", arsonistVoicelineVolumeArsonist.Value);
+            }
         }
 
         private static void ArsonistVoicelineVolume_Changed(object sender, EventArgs e)